+++ src/tgbot_expenses/utils/dollar_amount.py
@@ -1,17 +1,18 @@
 from src.tgbot_expenses.utils.currency_transfer import get_currency_rate
+from decimal import Decimal
 
 
-async def get_dollar_amount(account: str, amount: float) -> float:
+async def get_dollar_amount(account: str, amount: Decimal) -> Decimal:
     """
     Converts the given amount of a specified currency to its equivalent
     in US dollars.
 
     :param account: A string that represents the currency of the given amount.
     :type account: str
-    :param amount: A float that represents the amount of the given currency
+    :param amount: A Decimal that represents the amount of the given currency
                    to convert.
-    :type amount: float
-    :return: float
+    :type amount: Decimal
+    :return: Decimal
     """
     currencies = f'USD to {account.split(" ")[-1]}'
     currency_rate = 1 \
